# Perf: BN254 Miller loop