--- a/ecc/bn254/pairing.go
+++ b/ecc/bn254/pairing.go
@@ -203,7 +203,7 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 		l0.r0.MulByElement(&l0.r0, &p[k].Y)
 		l0.r1.MulByElement(&l0.r1, &p[k].X)
 
-		qProj[k].AddMixedStep(&l, &Q2)
+		qProj[k].AddMixedStepLineOnly(&l, &Q2)
 		l.r0.MulByElement(&l.r0, &p[k].Y)
 		l.r1.MulByElement(&l.r1, &p[k].X)
 		tmp.Mul034by034(&l.r0, &l.r1, &l.r2, &l0.r0, &l0.r1, &l0.r2)
@@ -289,4 +289,26 @@ func (p *g2Proj) AddMixedStep(evaluations *lineEvaluation, a *G2Affine) {
 	evaluations.r0.Set(&L)
 	evaluations.r1.Neg(&O)
 	evaluations.r2.Set(&J)
+}
+
+// AddMixedStepLineOnly computes the line through p in Homogenous projective
+// coordinates and a in affine coordinates. It does not compute the resulting
+// point p+a.
+func (p *g2Proj) AddMixedStepLineOnly(evaluations *lineEvaluation, a *G2Affine) {
+
+	// get some Element from our pool
+	var Y2Z1, X2Z1, O, L, t2, J fptower.E2
+	Y2Z1.Mul(&a.Y, &p.z)
+	O.Sub(&p.y, &Y2Z1)
+	X2Z1.Mul(&a.X, &p.z)
+	L.Sub(&p.x, &X2Z1)
+	t2.Mul(&L, &a.Y)
+	J.Mul(&a.X, &O).
+		Sub(&J, &t2)
+
+	// Line evaluation
+	evaluations.r0.Set(&L)
+	evaluations.r1.Neg(&O)
+	evaluations.r2.Set(&J)
+
 }