--- a/ecc/bw6-756/pairing.go
+++ b/ecc/bw6-756/pairing.go
@@ -191,13 +191,13 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 
 		// l_{p0,p1}(q)
 		pProj01[k].Set(&pProj1[k])
-		pProj01[k].AddMixedStep(&l01[k], &p0[k])
+		pProj01[k].addMixedStep(&l01[k], &p0[k])
 		l01[k].r1.Mul(&l01[k].r1, &q[k].X)
 		l01[k].r0.Mul(&l01[k].r0, &q[k].Y)
 
 		// l_{p0,-p1}(q)
 		pProj10[k].Neg(&pProj1[k])
-		pProj10[k].AddMixedStep(&l10[k], &p0[k])
+		pProj10[k].addMixedStep(&l10[k], &p0[k])
 		l10[k].r1.Mul(&l10[k].r1, &q[k].X)
 		l10[k].r0.Mul(&l10[k].r0, &q[k].Y)
 	}
@@ -213,7 +213,7 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 
 	// i = len(loopCounter) - 2
 	for k := 0; k < n; k++ {
-		pProj1[k].DoubleStep(&l0)
+		pProj1[k].doubleStep(&l0)
 		l0.r1.Mul(&l0.r1, &q[k].X)
 		l0.r0.Mul(&l0.r0, &q[k].Y)
 		result.MulBy034(&l0.r0, &l0.r1, &l0.r2)
@@ -227,14 +227,14 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 		j = loopCounter1[i]*3 + loopCounter0[i]
 
 		for k := 0; k < n; k++ {
-			pProj1[k].DoubleStep(&l0)
+			pProj1[k].doubleStep(&l0)
 			l0.r1.Mul(&l0.r1, &q[k].X)
 			l0.r0.Mul(&l0.r0, &q[k].Y)
 
 			switch j {
 			case -4:
 				tmp.Neg(&p01[k])
-				pProj1[k].AddMixedStep(&l, &tmp)
+				pProj1[k].addMixedStep(&l, &tmp)
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l01[k].r0, &l01[k].r1, &l01[k].r2)
@@ -242,13 +242,13 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 					Mul(&result, &ss)
 			case -3:
 				tmp.Neg(&p1[k])
-				pProj1[k].AddMixedStep(&l, &tmp)
+				pProj1[k].addMixedStep(&l, &tmp)
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l0.r0, &l0.r1, &l0.r2)
 				result.Mul(&result, &ss)
 			case -2:
-				pProj1[k].AddMixedStep(&l, &p10[k])
+				pProj1[k].addMixedStep(&l, &p10[k])
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l01[k].r0, &l01[k].r1, &l01[k].r2)
@@ -256,7 +256,7 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 					Mul(&result, &ss)
 			case -1:
 				tmp.Neg(&p0[k])
-				pProj1[k].AddMixedStep(&l, &tmp)
+				pProj1[k].addMixedStep(&l, &tmp)
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l0.r0, &l0.r1, &l0.r2)
@@ -264,27 +264,27 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 			case 0:
 				result.MulBy034(&l0.r0, &l0.r1, &l0.r2)
 			case 1:
-				pProj1[k].AddMixedStep(&l, &p0[k])
+				pProj1[k].addMixedStep(&l, &p0[k])
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l0.r0, &l0.r1, &l0.r2)
 				result.Mul(&result, &ss)
 			case 2:
 				tmp.Neg(&p10[k])
-				pProj1[k].AddMixedStep(&l, &tmp)
+				pProj1[k].addMixedStep(&l, &tmp)
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l01[k].r0, &l01[k].r1, &l01[k].r2)
 				result.MulBy034(&l0.r0, &l0.r1, &l0.r2).
 					Mul(&result, &ss)
 			case 3:
-				pProj1[k].AddMixedStep(&l, &p1[k])
+				pProj1[k].addMixedStep(&l, &p1[k])
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l0.r0, &l0.r1, &l0.r2)
 				result.Mul(&result, &ss)
 			case 4:
-				pProj1[k].AddMixedStep(&l, &p01[k])
+				pProj1[k].addMixedStep(&l, &p01[k])
 				l.r1.Mul(&l.r1, &q[k].X)
 				l.r0.Mul(&l.r0, &q[k].Y)
 				ss.Mul034By034(&l.r0, &l.r1, &l.r2, &l01[k].r0, &l01[k].r1, &l01[k].r2)
@@ -299,9 +299,9 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 	return result, nil
 }
 
-// DoubleStep doubles a point in Homogenous projective coordinates, and evaluates the line in Miller loop
+// doubleStep doubles a point in Homogenous projective coordinates, and evaluates the line in Miller loop
 // https://eprint.iacr.org/2013/722.pdf (Section 4.3)
-func (p *g1Proj) DoubleStep(evaluations *lineEvaluation) {
+func (p *g1Proj) doubleStep(evaluations *lineEvaluation) {
 
 	// get some Element from our pool
 	var t1, A, B, C, D, E, EE, F, G, H, I, J, K fp.Element
@@ -341,9 +341,9 @@ func (p *g1Proj) DoubleStep(evaluations *lineEvaluation) {
 	evaluations.r2.Set(&I)
 }
 
-// AddMixedStep point addition in Mixed Homogenous projective and Affine coordinates
+// addMixedStep point addition in Mixed Homogenous projective and Affine coordinates
 // https://eprint.iacr.org/2013/722.pdf (Section 4.3)
-func (p *g1Proj) AddMixedStep(evaluations *lineEvaluation, a *G1Affine) {
+func (p *g1Proj) addMixedStep(evaluations *lineEvaluation, a *G1Affine) {
 
 	// get some Element from our pool
 	var Y2Z1, X2Z1, O, L, C, D, E, F, G, H, t0, t1, t2, J fp.Element