Actually we don't need to compute a full `AddMixedStep` [here](https://github.com/ConsenSys/gnark-crypto/blob/8f7ca09273c24ed9465043566906cbecf5dcee91/ecc/bn254/pairing.go#L206). We just need the line that goes through `qProj` and `Q2` but not the new `qProj` (`=qProj+Q2`) as it won't be used after that. 
This PR introduces the private function `addMixedStepLineOnly` that only computes the line and not the resulting point, and makes `AddMixedStep` and `DoubleStep` private for all curves.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkMillerLoop-2               210243        209380        -0.41%
BenchmarkMultiMiller/2_pairs-2      353040        352036        -0.28%
BenchmarkMultiMiller/3_pairs-2      495552        493883        -0.34%
BenchmarkMultiMiller/4_pairs-2      637746        635996        -0.27%
BenchmarkMultiMiller/5_pairs-2      779299        777806        -0.19%
BenchmarkMultiMiller/6_pairs-2      921408        919538        -0.20%
BenchmarkMultiMiller/7_pairs-2      1063256       1061098       -0.20%
BenchmarkMultiMiller/8_pairs-2      1206537       1202637       -0.32%
BenchmarkMultiMiller/9_pairs-2      1346727       1342862       -0.29%
BenchmarkMultiMiller/10_pairs-2     1488339       1485146       -0.21%
```